model-api/requirements.txt
flask==2.3.3
tensorflow==2.12.0
keras==2.12.0
pillow==9.5.0
numpy==1.23.5